refactorization of the decorator to make it a general case
