+++ contentcuration/contentcuration/utils/channel.py
@@ -1,8 +1,3 @@
-import functools
-import math
-import random
-import time
-
 from django.core.cache import cache
 from django.db.models import Count
 from django.db.models import Sum
@@ -12,85 +7,16 @@ from contentcuration.models import Channel
 from contentcuration.models import ContentNode
 from contentcuration.models import FileCTE
 from contentcuration.models import User
+from contentcuration.utils.cache import cache_stampede
 
 CACHE_CHANNEL_KEY = "channel_metadata_{}"
 
 
-def cache_stampede(expire, beta=1):
-    """Cache decorator with cache stampede protection.
-    Based on http://www.vldb.org/pvldb/vol8/p886-vattani.pdf (research by
-    Vattani, A.; Chierichetti, F.; Lowenstein, K. (2015), Optimal Probabilistic
-    Cache Stampede Prevention, VLDB, pp. 886-897, ISSN 2150-8097) and the
-    Python implementation at
-    https://github.com/grantjenks/python-diskcache/blob/master/diskcache/recipes.py#L315
-
-    The cache stampede problem (also called dog-piling, cache miss storm,
-    or cache choking) is a situation that occurs when a popular cache item
-    expires, leading to multiple requests seeing a cache miss and
-    regenerating that same item at the same time
-
-    This  decorator implements cache stampede protection through
-    early recomputation. Early recomputation of function results will occur
-    probabilistically before expiration in a background thread of
-    execution.
-
-    :param float expire: seconds until arguments expire
-    :param int beta: the parameter beta can be set to a value greater than 1 to
-                     favor earlier recomputations and further reduce stampedes but
-                     the paper authors show that setting beta=1 works well in practice
-    :return: callable decorator
-    """
-
-    def decorator(func):
-        def timer(*args, **kwargs):
-            "Time execution of `func` and return result and time delta."
-            start = time.time()
-            result = func(*args, **kwargs)
-            delta = time.time() - start
-            # The variable delta represents the time to recompute the value
-            # and is used to scale the probability distribution appropriately.
-            return result, delta
-
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            channel_id = args[0]
-            key = CACHE_CHANNEL_KEY.format(channel_id)
-            cached = cache.get(key)
-            if cached is not None:
-                metadata = cached["METADATA"]
-                if cached["CALCULATING"]:
-                    return metadata
-                expire_time = cached["EXPIRE"]
-                now = time.time()
-                ttl = expire_time - now
-                delta = cached["DELTA"]
-                if (-delta * math.log(random.random())) < ttl:
-                    return metadata  # Cache hit.
-            metadata, delta = timer(*args, *kwargs)
-            cached_info = {
-                "CALCULATING": False,
-                "METADATA": metadata,
-                "DELTA": delta,
-                "EXPIRE": time.time() + expire,
-            }
-            cache.set(key, cached_info, timeout=None)
-
-        return wrapper
-
-    return decorator
-
-
 @cache_stampede(expire=3600)
-def calculate_channel_metadata(channel_id=None, tree_id=None):
-    if channel_id is None and tree_id is None:
+def calculate_channel_metadata(key, channel_id=None, tree_id=None):
+    if key is None:
         return  # this is an error, it should not happen, but just in case
 
-    if channel_id is None:
-        channel_id = Channel.objects.filter(main_tree__tree_id=tree_id).values_list(
-            "id", flat=True
-        )[0]
-
-    key = CACHE_CHANNEL_KEY.format(channel_id)
     cached_info = cache.get(key)
     if cached_info is not None:
         if cached_info["CALCULATING"]:
